add note about refetch variables
add brief note to clear any variables that were set before for any query. This is helpful when you want to omit a variable from the query that was set before. e.g in case of filters you might want to clear previous filter variable but by not passing that variable will result in picking the old value.